Fluid

HeatTransferFluids.Fluid — Type

Fluid{
    T1<:Unitful.Density,
    T2<:Unitful.Velocity,
    T3<:Unitful.DynamicViscosity,
    T4<:SpecificHeatCapacity,
    T5<:ThermalConductivity,
    T6<:Unitful.Temperature,
}

A fluid with the important physical properties to be used in heat transfer calculations.

Fields

density::T1: the density of the fluid

velocity::T2: the velocity of the fluid

viscosity::T3: the viscosity of the fluid

heat_capacity::T4: the specific heat capacity of the fluid

thermal_conductivity::T5: the thermal conductivity of the fluid

temperature::T6: the temperature of the fluid

HeatTransferFluids.Water — Function

Water

Returns a Fluid with the properties of water close to room temperature.

Keywords

velocity::Unitful.Velocity: the velocity of the water

temperature::Unitful.Temperature: the temperature of the water

Returns

Fluid: the fluid with the properties of water close to room temperature at the given velocity and temperature