Heat transfer

HeatTransferFluids.heat_transfer — Function

heat_transfer(f::Fluid, t::Tube)

Computes the heat transfer coefficient of a fluidflowing through atube`.

Arguments

f::Fluid: the fluid flowing through the tube

t::Tube: the tube through which the fluid flows

Returns

Unitful.Quantity: the heat transfer coefficient of the fluid flowing through the tube