Structure

HeatTransferFluids.Structure — Type

Structure

A structure is an element through which a fluid flows.

HeatTransferFluids.Tube — Type

Tube

A tube with a diameter and length through which a fluid flows.

Fields

diameter::Unitful.Length: the diameter of the tube

length::Unitful.Length: the length of the tube

HeatTransferFluids.Coil — Type

Coil

A tube in the shape of a coil with a diameter and pitch.

Fields

tube::Tube: the tube that is coiled

diameter::Unitful.Length: the diameter of the coil

pitch::Unitful.Length: the pitch of the coil

HeatTransferFluids.Valve — Type

Valve

A valve parametrized by a flow rate and flow factor.

Fields

flow_rate::VolumeFlow: the flow rate of the valve

flow_factor::VolumeFlow: the flow factor of the valve